refactor(components): refine taxonomy (browse split, telemetry by scope)
Follow-up to the feature-folder reorg, per review:

- Revert AnthropicMark/GeminiMark back to *Logo — the three brand marks are
  intentionally NOT uniform (GitHub ships a 'mark', the AI brands are logos).
- Add browse/ for the lens-list page: orchestration (LensListClient,
  LensIndexShell, LensSectionNav) plus its toolbar controls (LensSortControl,
  LensUsageSwitch, LensSearchDialog+telemetry). lens/ now holds only reusable
  lens atoms (LensCard, SpecialtyBadges). Sort is not a filter, so it lands in
  browse/ rather than lens-filters/.
- PromoBanner → ui/promo-banner.tsx: it's a generic dismissible campaign
  banner with no price coupling (and currently no consumers).
- Merge poster/ under share/poster/ — a poster is a share artifact.
- Telemetry by scope: feature-specific telemetry colocates with its feature
  (CompareTelemetry→compare/, LensDetailTelemetry→lens/); telemetry/ keeps only
  the global trackers (AssetTelemetry, OutboundLinkTelemetry, PwaLaunchTelemetry).
- price/ owns all price/purchase rendering wherever it appears (LensCardPrice
  included), resolving the lens-vs-price overlap by domain ownership.

tsc clean, eslint clean, 383 tests pass.
