+++ src/app/[locale]/lenses/[mount]/(browse)/browse/page.tsx
@@ -3,7 +3,7 @@ import type { Metadata } from "next";
 import { getTranslations, setRequestLocale } from "next-intl/server";
 import { getLensesByMount } from "@/lib/lens/data";
 import { urlSegmentToMount } from "@/lib/mount";
-import LensListClient from "@/components/lens/LensListClient";
+import LensListClient from "@/components/browse/LensListClient";
 import LensesLoading from "../loading";
 import { buildAlternates, defaultOgImages } from "@/lib/seo";
 import { notFound } from "next/navigation";
