+++ src/app/[locale]/lenses/[mount]/(browse)/collections/page.tsx
@@ -20,7 +20,7 @@ import { getLensesByMount } from "@/lib/lens/data";
 import { urlSegmentToMount, mountHasCollections } from "@/lib/mount";
 import { buildAlternates, defaultOgImages } from "@/lib/seo";
 import { ACTION_PRIMARY_CLS } from "@/config/ui-tokens";
-import LensIndexShell from "@/components/lens/LensIndexShell";
+import LensIndexShell from "@/components/browse/LensIndexShell";
 import CollectionChipRail from "@/components/collection/CollectionChipRail";
 
 type Params = Promise<{ locale: string; mount: string }>;
