+++ src/components/browse/LensSearchDialog.telemetry.ts
