+++ src/components/browse/LensSectionNav.tsx
