+++ src/components/compare/CompareTelemetry.tsx
