+++ src/components/lens/LensDetailTelemetry.hooks.ts
