+++ src/components/lens/LensDetailTelemetry.tsx
