+++ src/components/share/poster/PosterFocalRuler.tsx
